imgsample foundations

+++ test.h
@@ -1,2 +1,2 @@
 double myadder(double x, double y);
-void add_array(int row_count, int column_count, double *a, double *b, double *result);
+void add_array(size_t row_count, size_t column_count, double *a, double *b, double *result);
